2D

+++ cute_kernels/kernels/swiglu_unchunked/triton_implementation/kernels_backward.py
@@ -22,7 +22,7 @@ def swiglu_unchunked_backward_triton_kernel(
     up = tl.load(up_ptrs, mask=mask_bh)
 
     gate_ptrs = up_ptrs + (H >> 1)
-    gate = tl.load(gate_ptrs, mask=mask_bh)
+    gate = tl.load(gate_ptrs, mask=mask_bh).to(tl.float32)
 
     output_grad_ptrs = output_grad_ptr + indices_b[:, None] * half_H + indices_h[None, :]
     output_grad = tl.load(output_grad_ptrs, mask=mask_bh)
